feat(caritas-theme): remove extension, move i18n to core
