--- src/extensions/cypress/cypress.json
@@ -1,12 +1,0 @@
-{
-	"e2e": {
-		"excludeSpecPattern": [
-			"cypress/e2e/**/registration/base.cy.ts",
-			"cypress/e2e/**/registration/consultingType.cy.ts",
-			"cypress/e2e/**/registration/topic.cy.ts",
-			"cypress/e2e/**/messages/attachments/informal.cy.ts",
-			"cypress/e2e/**/login.cy.ts"
-		],
-		"specPattern": ["src/extensions/cypress/e2e/**/*.cy.ts"]
-	}
-}
