--- src/extensions/resources/i18n/de/agency.json
@@ -1,3 +1,0 @@
-{
-	"agency": {}
-}
