--- src/extensions/resources/i18n/de/languages.json
@@ -1,4 +1,0 @@
-{
-	"de": "(DE) Deutsch",
-	"en": "(EN) Englisch"
-}
