--- src/extensions/resources/i18n/en/agency.json
@@ -1,3 +1,0 @@
-{
-	"agency": {}
-}
