--- src/extensions/resources/i18n/en/languages.json
@@ -1,4 +1,0 @@
-{
-	"de": "(DE) German",
-	"en": "(EN) English"
-}
