+++ src/resources/i18n/de/languages.json
@@ -1,3 +1,4 @@
 {
-	"de": "(DE) Deutsch"
+	"de": "(DE) Deutsch",
+	"en": "(EN) Englisch"
 }
