+++ src/resources/scripts/config.ts
@@ -1,5 +1,10 @@
 import deAgency from '../i18n/de/agency.json';
 import deConsultingTypes from '../i18n/de/consultingTypes.json';
+import deLanguages from '../i18n/de/languages.json';
+import en from '../i18n/en/common.json';
+import enAgency from '../i18n/en/agency.json';
+import enConsultingTypes from '../i18n/en/consultingTypes.json';
+import enLanguages from '../i18n/en/languages.json';
 import { AppConfigInterface } from '../../globalState/interfaces';
 import {
 	OVERLAY_RELEASE_NOTE,
@@ -9,6 +14,14 @@ import {
 export const uiUrl = window.location.origin;
 
 export const APP_PATH = 'app';
+
+export const routePathNames = {
+	root: '/',
+	login: '/login',
+	termsAndConditions: '/nutzungsbedingungen',
+	imprint: '/impressum',
+	privacy: '/datenschutz'
+};
 
 export const config: AppConfigInterface = {
 	budibaseUrl: '',
@@ -32,7 +45,7 @@ export const config: AppConfigInterface = {
 		}
 	},
 	urls: {
-		chatScheduleUrl: 'https://www.caritas.de/onlineberatung',
+		chatScheduleUrl: uiUrl + '/registration',
 		consultantVideoConference:
 			'/consultant/videoberatung/:type/:appointmentId',
 		error401: uiUrl + '/error.401.html',
@@ -40,14 +53,14 @@ export const config: AppConfigInterface = {
 		error500: uiUrl + '/error.500.html',
 		finishedAnonymousChatRedirect:
 			'https://www.caritas.de/hilfeundberatung/hilfeundberatung',
-		home: 'https://www.caritas.de',
-		landingpage: '/login',
+		home: uiUrl + '/registration',
+		landingpage: '/registration',
 		releases: uiUrl + '/releases',
 		redirectToApp: uiUrl + '/' + APP_PATH,
 		registration: uiUrl + '/registration',
-		toEntry: uiUrl + '/',
+		toEntry: uiUrl + '/registration',
 		toLogin: uiUrl + '/login',
-		toRegistration: 'https://www.caritas.de/onlineberatung',
+		toRegistration: uiUrl + '/registration',
 		videoCall: '/videoanruf/:domain/:jwt/:video?/:username?/:e2e?',
 		videoConference: '/videoberatung/:type/:appointmentId'
 	},
@@ -57,7 +70,7 @@ export const config: AppConfigInterface = {
 			showCreationDate: false
 		}
 	},
-	postcodeFallbackUrl: '{url}',
+	postcodeFallbackUrl: '{url}{postcode}/',
 	jitsi: {
 		/**
 		 * Enable WebRTC Encoded Transform as an alternative to insertable streams.
@@ -75,14 +88,19 @@ export const config: AppConfigInterface = {
 	},
 	legalLinks: [
 		{
-			url: 'https://www.caritas.de/impressum',
+			url: uiUrl + routePathNames.imprint,
 			label: 'login.legal.infoText.impressum'
 		},
 		{
-			url: 'https://www.caritas.de/hilfeundberatung/onlineberatung/datenschutz',
+			url: uiUrl + routePathNames.privacy,
 			label: 'login.legal.infoText.dataprotection',
 			registration: true
 		}
+		// {
+		// 	url: routePathNames.termsAndConditions,
+		// 	label: 'login.legal.infoText.termsAndConditions',
+		// 	registration: true
+		// }
 	],
 	welcomeScreen: {
 		consultingType: {
@@ -91,7 +109,7 @@ export const config: AppConfigInterface = {
 	},
 	registration: {
 		consultingTypeDefaults: {
-			autoSelectPostcode: true,
+			autoSelectPostcode: false,
 			autoSelectAgency: false
 		}
 	},
@@ -312,15 +330,33 @@ export const config: AppConfigInterface = {
 		'zu'
 	],
 	i18n: {
-		supportedLngs: ['de', 'de@informal'],
-		preload: ['de', 'de@informal'],
+		supportedLngs: ['en', 'de'],
+		fallbackLng: {
+			en: ['de'],
+			en_informal: ['en', 'de_informal', 'de']
+		},
 		resources: {
 			de: {
+				languages: {
+					...deLanguages
+				},
 				consultingTypes: {
 					...deConsultingTypes
 				},
 				agencies: {
 					...deAgency
+				}
+			},
+			en: {
+				common: [en],
+				consultingTypes: {
+					...enConsultingTypes
+				},
+				agencies: {
+					...enAgency
+				},
+				languages: {
+					...enLanguages
 				}
 			}
 		}
